+++ internal/models/rd/client/role_permission.go
@@ -2,8 +2,8 @@ package rdClient
 
 // RolePermission 角色权限表
 type RolePermission struct {
-	RoleID       uint `json:"role_id"       gorm:"not null;"` // 角色 ID，普通字段
-	PermissionID uint `json:"permission_id" gorm:"not null;"` // 权限 ID，普通字段
+	RoleID       uint64 `json:"role_id"       gorm:"not null;"` // 角色 ID，普通字段
+	PermissionID uint64 `json:"permission_id" gorm:"not null;"` // 权限 ID，普通字段
 }
 
 func (c *RolePermission) TableName() string {
